Move into own variable files
