+++ policyengine_us/variables/gov/states/ca/tax/income/credits/earned_income/ca_eitc.py
@@ -130,52 +130,3 @@ class ca_eitc(Variable):
             eitc_along_second_phase_out,
             amount,
         )
-
-
-class ca_eitc_second_phase_out_start(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "CalEITC second phase-out start"
-    unit = USD
-    documentation = "California begins secondarily phasing out the CalEITC at this earnings level."
-    definition_period = YEAR
-
-    def formula(tax_unit, period, parameters):
-        maximum = tax_unit("ca_eitc_maximum", period)
-        ca_eitc = parameters(
-            period
-        ).gov.states.ca.tax.income.credits.earned_income
-        count_children = tax_unit("eitc_child_count", period)
-        # Apply the adjustment factor to the phase-out rate.
-        phase_out_rate = (
-            ca_eitc.phase_out.rate.calc(count_children)
-            * ca_eitc.adjustment.factor
-        )
-        # Multiply and divide the second phase-out start by the adjustment factors.
-        eitc_at_which_second_phase_out_starts = (
-            ca_eitc.phase_out.final.start.calc(count_children)
-            * ca_eitc.adjustment.factor
-            / ca_eitc.adjustment.divisor
-        )
-        eitc_difference = maximum - eitc_at_which_second_phase_out_starts
-        earnings_to_cover_eitc_difference = eitc_difference / phase_out_rate
-        return (
-            earnings_to_cover_eitc_difference
-            + ca_eitc.phase_out.start.calc(count_children)
-        )
-
-
-class ca_eitc_maximum(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "CalEITC maximum"
-    unit = USD
-    documentation = "California provides an earned income tax credit up to this value, based on the earned income amounts and phase-in rates, which vary with number of qualifying children."
-    definition_period = YEAR
-
-    def formula(tax_unit, period, parameters):
-        branch = get_ca_eitc_branch(tax_unit, period, parameters)
-        adjustment = parameters(
-            period
-        ).gov.states.ca.tax.income.credits.earned_income.adjustment.factor
-        return branch.calculate("eitc_maximum", period) * adjustment
